Remove `TranslatorModuleContext`
