+++ Sources/WasmKit/Translator.swift
@@ -111,94 +111,6 @@ extension InternalInstance: TranslatorContext {
     }
 }
 
-
-struct TranslatorModuleContext {
-    internal let imports: ModuleImports
-    internal let typeSection: [FunctionType]
-    internal let functionTypeIndices: [TypeIndex]
-    internal let globalTypes: [GlobalType]
-    internal let memoryTypes: [MemoryType]
-    internal let tableTypes: [TableType]
-
-    init(
-        typeSection: [FunctionType],
-        importSection: [Import],
-        functionSection: [TypeIndex],
-        globals: [GlobalType],
-        memories: [MemoryType],
-        tables: [TableType]
-    ) {
-        var functionTypeIndices: [TypeIndex] = []
-        var globalTypes: [GlobalType] = []
-        var memoryTypes: [MemoryType] = []
-        var tableTypes: [TableType] = []
-
-        self.imports = ModuleImports.build(
-            from: importSection,
-            functionTypeIndices: &functionTypeIndices,
-            globalTypes: &globalTypes,
-            memoryTypes: &memoryTypes,
-            tableTypes: &tableTypes
-        )
-
-        self.typeSection = typeSection
-        self.functionTypeIndices = functionTypeIndices + functionSection
-        for global in globals {
-            globalTypes.append(global)
-        }
-        self.globalTypes = globalTypes
-        self.memoryTypes = memoryTypes + memories
-        self.tableTypes = tableTypes + tables
-    }
-
-    func resolveType(_ index: TypeIndex) throws -> FunctionType {
-        guard Int(index) < typeSection.count else {
-            throw TranslationError("Type index \(index) is out of range")
-        }
-        return typeSection[Int(index)]
-    }
-    func resolveBlockType(_ blockType: BlockType) throws -> FunctionType {
-        try FunctionType(blockType: blockType, typeSection: typeSection)
-    }
-    func functionType(_ index: FunctionIndex, interner: Interner<FunctionType>) throws -> FunctionType {
-        guard Int(index) < functionTypeIndices.count else {
-            throw TranslationError("Function index \(index) is out of range")
-        }
-        let typeIndex = functionTypeIndices[Int(index)]
-        return try resolveType(typeIndex)
-    }
-    func globalType(_ index: GlobalIndex) throws -> ValueType {
-        guard Int(index) < globalTypes.count else {
-            throw TranslationError("Global index \(index) is out of range")
-        }
-        return self.globalTypes[Int(index)].valueType
-    }
-    func isMemory64(memoryIndex index: MemoryIndex) throws -> Bool {
-        guard Int(index) < memoryTypes.count else {
-            throw TranslationError("Memory index \(index) is out of range")
-        }
-        return self.memoryTypes[Int(index)].isMemory64
-    }
-    func isMemory64(tableIndex index: TableIndex) throws -> Bool {
-        guard Int(index) < tableTypes.count else {
-            throw TranslationError("Table index \(index) is out of range")
-        }
-        return self.tableTypes[Int(index)].limits.isMemory64
-    }
-    func elementType(_ index: TableIndex) throws -> ReferenceType {
-        guard Int(index) < tableTypes.count else {
-            throw TranslationError("Table index \(index) is out of range")
-        }
-        return self.tableTypes[Int(index)].elementType
-    }
-    func resolveCallee(_ index: FunctionIndex) -> InternalFunction? {
-        return nil
-    }
-    func resolveGlobal(_ index: GlobalIndex) -> InternalGlobal? {
-        return nil
-    }
-}
-
 fileprivate struct MetaProgramCounter {
     let offsetFromHead: Int
 }
